Finished updating messages to properly support dynamic message lengths. Added MS Visual Studio files to the svn:ignore properties.

+++ openjaus/branches/OpenJAUSv3.2/libjaus/src/message/experimental/reportPayloadInterfaceMessage.c
@@ -812,7 +812,7 @@ static int dataSize(ReportPayloadInterfaceMessage message)
 		if(tempUShort != 0)
 		{
 			tempString = jausGetInformationInterfaceEnumeration(message->jausPayloadInterface, identifierString, &successFlag);
-			index += strlen(tempString);
+			index += (int) strlen(tempString);
 		}
 
 		// get/pack HMI recommendation, if in PV
